requirements.txt
python-telegram-bot==20.8
yt-dlp==2024.08.06
uvloop==0.19.0
